Time stop, new ingredient art, game ending/credits
Time stop overlay happens when the game is paused or when you're in a
spell circle. Hourglass is placeholder. Ingredient prefabs are plugged
into creature manager, and the inventory script has had variables added
to track them (plus two numbers in the UI for them), but don't do
anything related to spells yet.