Tutorial pop up info
Put in the system for all the messages that will pop up to guid the
player through the UI. Still needs to be railroaded a bit more because
it can't function out of order. Still some bugs here and there, but the
messages are all present and legible. It's run out of a tutorial manager
which awkwardly talks to the dialogue manager to advance.